deployment/scraper/extract_city/requirements.txt
locationtagger
spacy
pandas
psycopg2-binary